+++ noodles-cram/CHANGELOG.md
@@ -6,7 +6,8 @@
 
   * cram/codecs: Add fqzcomp encoder.
 
-  * cram/codecs/name_tokenizer/encode: Add match tokens.
+  * cram/codecs/name_tokenizer/encode: Add `DIGITS0` and `MATCH`
+    tokens.
 
 ### Fixed